# docs/openapi.yaml
openapi: 3.0.1
info:
  title: MOS (Meal and Reminder System) API
  description: API for managing meal requirements, reminders, and user authentication
  version: 1.0.0
  contact:
    name: MOS Development Team
    email: dev@mos.example.com
servers:
- url: http://localhost:8080
  description: Development server
- url: https://api.mos.example.com
  description: Production server
paths:
...
  /api/users/{userId}/reminders:
    get:
      tags:
      - reminders
      summary: Get user reminders
      description: Retrieve all reminders for a specific user
      operationId: getReminders
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderDto"
        "403":
          description: Forbidden - User doesn't have permission to view reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to view reminders"
    post:
      tags:
      - reminders
      summary: Create a reminder
      description: "Create a new reminder with time, category, and optional note."
      operationId: createReminder
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderRequest"
        required: true
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderDto"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
                  message:
                    type: string
                    example: "time must be now or future (±5m)"
        "403":
          description: Forbidden - User doesn't have permission to create reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to create reminders"
  /api/users/{userId}/reminders/{reminderId}:
    put:
      tags:
      - reminders
      summary: Update a reminder
      description: Update an existing reminder with new time, category, or note
      operationId: updateReminder
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      - name: reminderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Reminder ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReminderRequest"
            examples:
              updateOnceReminder:
                summary: Update a one-time reminder
                value:
                  type: "once"
                  category: "medicin"
                  dateTime: "2025-09-27T14:30:00Z"
                  note: "Updated medication reminder"
              updateRecurringReminder:
                summary: Update a recurring reminder
                value:
                  type: "recurring"
                  category: "måltider"
                  days: ["Mon", "Wed", "Fri"]
                  times: ["08:00", "12:00", "18:00"]
                  note: "Updated meal reminders"
              partialUpdate:
                summary: Partial update (only note)
                value:
                  note: "Just updating the note"
      responses:
        "200":
          description: Reminder updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderDto"
              example:
                id: "550e8400-e29b-41d4-a716-446655440001"
                time: "2025-09-27T14:30:00Z"
                category: "MEDICATION"
                note: "Updated medication reminder"
                createdAt: "2025-09-26T12:00:00Z"
                type: "once"
                recurrence: null
        "400":
          description: Invalid input or reminder not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
                  message:
                    type: string
                    example: "Reminder not found for user"
        "403":
          description: Forbidden - User doesn't have permission to update reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to update reminders"
        "404":
          description: Reminder not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
                  message:
                    type: string
                    example: "Reminder not found"
    delete:
      tags:
      - reminders
      summary: Delete a reminder
      description: Delete a specific reminder by ID
      operationId: deleteReminder
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      - name: reminderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Reminder ID
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden - User doesn't have permission to delete reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to delete reminders"
  /api/users/{userId}/meal-requirements:
    get:
      tags:
      - meal-requirements
      summary: Get user meal requirements
      description: Retrieve all meal requirements for a specific user
      operationId: getMealRequirements
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealRequirementsResponse"
        "403":
          description: Forbidden - User doesn't have permission to view meal requirements
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to view meal requirements"
    post:
      tags:
      - meal-requirements
      summary: Set meal requirements
      description: Set meal requirements for a user. Replaces all existing requirements.
      operationId: setMealRequirements
      security:
      - bearerAuth: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealRequirementsRequest"
        required: true
      responses:
        "201":
          description: Meal requirements set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealRequirementsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
                  message:
                    type: string
                    example: "Validation failed"
        "403":
          description: Forbidden - User doesn't have permission to manage meal requirements
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "You don't have permission to manage meal requirements"

  # User Management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    CreateReminderRequest:
      type: object
      required:
        - type
        - category
        - dateTime
      properties:
        type:
          type: string
          description: Reminder type
          example: "måltider"
        category:
          type: string
          enum: [MEDICATION, MEAL]
          description: Reminder category
          example: "MEAL"
        dateTime:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2025-09-22T12:00:00Z"
        note:
          type: string
          description: Optional note
          example: "Lunch reminder"
    UpdateReminderRequest:
      type: object
      description: Request body for updating a reminder (all fields optional for partial updates)
      properties:
        type:
          type: string
          enum: ["once", "recurring"]
          description: Reminder type
          example: "once"
        category:
          type: string
          description: Reminder category (Swedish or English)
          example: "medicin"
        dateTime:
          type: string
          format: date-time
          description: Reminder date and time (for once type)
          example: "2025-09-27T14:30:00Z"
        days:
          type: array
          items:
            type: string
          description: Days of the week for recurring reminders
          example: ["Mon", "Wed", "Fri"]
        times:
          type: array
          items:
            type: string
          description: Times for recurring reminders
          example: ["08:00", "12:00", "18:00"]
        note:
          type: string
          description: Optional note
          example: "Updated reminder note"
    ReminderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Reminder ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        time:
          type: string
          format: date-time
          description: Reminder time
          example: "2025-09-22T12:00:00Z"
        category:
          type: string
          enum: [MEDICATION, MEAL]
          description: Reminder category
          example: "MEAL"
        note:
          type: string
          description: Reminder note
          example: "Lunch reminder"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-09-22T09:13:56.743059+00"

    # Meal requirement schemas
    MealRequirementsRequest:
      type: object
      required:
        - requirements
      properties:
        requirements:
          type: array
          items:
            type: string
          example: ["Laktosfri", "Diabetesanpassad"]
    MealRequirementsResponse:
      type: object
      properties:
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/MealRequirementDto'
    MealRequirementDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        requirement:
          type: string
        createdAt:
          type: string
          format: date-time

    # User Management schemas
...